V1.3.0
Solved https://github.com/5axes/Calibration-Shapes/issues/1
