+++ plugin.json
@@ -1,7 +1,7 @@
 {
     "name": "Calibration Shapes",
     "author": "5axes",
-    "version": "1.2.6",
+    "version": "1.3.0",
     "description": "Adds test parts and simple shapes to the build plate",
     "api": 7,
     "supported_sdk_versions": ["7.0.0", "7.1.0", "7.2.0", "7.3.0", "7.4.0", "7.5.0"],
